Move Countries class to components folder
